requirements.txt
google-genai>=1.30.0
google-generativeai>=0.8.5
streamlit>=1.45.0
numpy>=1.26.4
faiss-cpu>=1.8.0.post1
librosa>=0.10.2
soundfile>=0.12.1
moviepy>=2.1.1
python-dotenv>=1.0.1

